+++ R/sensor_functions.R
@@ -804,12 +804,12 @@ identify_gaps <- function(db, participant_id = NULL, min_gap = 60, sensor = "Acc
 #'   the analysis. This function adds known gaps to data as "measurements", thereby allowing easier
 #'   calculations for, for example, finding the duration. For instance, consider a participant spent
 #'   30 minutes walking. However, if it is known there is gap of 15 minutes in this interval, we
-#'   should somehow account for it. \code{add_gaps} accounts for this by adding the gap data to
+#'   should somehow account for it. `add_gaps` accounts for this by adding the gap data to
 #'   sensors data by splitting intervals where gaps occur.
 #'
 #' @details In the example of 30 minutes walking where a 15 minute gap occurred (say after 5
-#'   minutes), \code{add_gaps} adds two rows: one after 5 minutes of the start of the interval
-#'   indicating the start of the gap(if needed containing values from \code{fill}), and one after 20
+#'   minutes), `add_gaps()` adds two rows: one after 5 minutes of the start of the interval
+#'   indicating the start of the gap(if needed containing values from `fill`), and one after 20
 #'   minutes of the start of the interval signalling the walking activity. Then, when calculating
 #'   time differences between subsequent measurements, the gap period is appropriately accounted
 #'   for. Note that if multiple measurements occurred before the gap, they will both be continued
@@ -817,21 +817,20 @@ identify_gaps <- function(db, participant_id = NULL, min_gap = 60, sensor = "Acc
 #'
 #' @inheritSection identify_gaps Warning
 #'
-#' @param data A data frame containing the data. See [get_data()] for
-#'   retrieving activity data from an mpathsenser database.
-#' @param gaps A data frame (extension) containing the gap data. See
-#'   \link[mpathsenser]{identify_gaps} for retrieving gap data from an mpathsenser database. It
-#'   should at least contain the columns \code{from} and \code{to} (both in a date-time format), as
-#'   well as any specified columns in \code{by}.
+#' @param data A data frame containing the data. See [get_data()] for retrieving data from an
+#'   mpathsenser database.
+#' @param gaps A data frame (extension) containing the gap data. See [identify_gaps()] for
+#'   retrieving gap data from an mpathsenser database. It should at least contain the columns `from`
+#'   and `to` (both in a date-time format), as well as any specified columns in `by`.
 #' @param by A character vector indicating the variable(s) to match by, typically the participant
-#'   IDs. If NULL, the default, \code{*_join()} will perform a natural join, using all variables in
-#'   common across \code{x} and \code{y}.
+#'   IDs. If NULL, the default, `*_join()` will perform a natural join, using all variables in
+#'   common across `x and `y`.
 #' @param continue Whether to continue the measurement(s) prior to the gap once the gap ends.
 #' @param fill A named list of the columns to fill with default values for the extra measurements
 #'   that are added because of the gaps.
 #'
-#' @seealso [identify_gaps()] for finding gaps in the sampling;
-#'   [link_gaps()] for linking gaps to ESM data, analogous to [link()].
+#' @seealso [identify_gaps()] for finding gaps in the sampling; [link_gaps()] for linking gaps to
+#'   ESM data, analogous to [link()].
 #'
 #' @returns A tibble containing the data and the added gaps.
 #' @export
@@ -893,7 +892,8 @@ add_gaps <- function(data, gaps, by = NULL, continue = FALSE, fill = NULL) {
     gaps <- dplyr::semi_join(gaps, data, by = rlang::as_name(rlang::enquo(by)))
   }
 
-  # If we don't want to continue the previous measurement after the gap, we can
+  # If we don't want to continue the previous measurement after the gap, we can simply add the
+  # gaps to the data and sort
   if (!continue) {
     gaps <- gaps %>%
       select({{ by }}, time = "from") %>%